@@ -26,6 +26,7 @@
 		"test": "vue-cli-service test:unit"
 	},
 	"dependencies": {
+		"@vue/compat": "^3.2.0",
 		"axios": "^0.27.2",
 		"base64-js": "^1.5.1",
 		"bootstrap": "^4.6.1",
@@ -40,10 +41,10 @@
 		"pako": "^2.0.4",
 		"reflect-metadata": "^0.1.13",
 		"tippy.js": "^6.3.7",
-		"vue": "^2.7.13",
-		"vue-select": "^3.18.3",
+		"vue": "^3.2.0",
+		"vue-select": "^4.0.0-beta.5",
 		"vuedraggable": "github:FelixRilling/Vue.Draggable#fb1243deb651bc3d97f9df71834d0aeab98e4dd2",
-		"vuex": "^3.6.2"
+		"vuex": "^4.1.0"
 	},
 	"devDependencies": {
 		"@rushstack/eslint-patch": "^1.1.4",
@@ -60,9 +61,10 @@
 		"@vue/cli-plugin-unit-jest": "~5.0.6",
 		"@vue/cli-plugin-vuex": "~5.0.6",
 		"@vue/cli-service": "~5.0.6",
+		"@vue/compiler-sfc": "^3.2.0",
 		"@vue/eslint-config-typescript": "^11.0.0",
-		"@vue/test-utils": "^1.1.3",
-		"@vue/vue2-jest": "^27.0.0-alpha.2",
+		"@vue/test-utils": "^2.2.1",
+		"@vue/vue3-jest": "^27.0.0-alpha.1",
 		"babel-jest": "^27.0.6",
 		"eslint": "^8.19.0",
 		"eslint-config-prettier": "^8.5.0",
@@ -74,7 +76,6 @@
 		"sass-loader": "^13.0.2",
 		"ts-jest": "^27.0.4",
 		"ts-mockito": "^2.6.1",
-		"typescript": "~4.7.4",
-		"vue-template-compiler": "^2.7.13"
+		"typescript": "~4.7.4"
 	}
 }
