@@ -1,5 +1,4 @@
-import Vue from "vue";
-import Vuex, { Store } from "vuex";
+import { createStore } from "vuex";
 import { DEVELOPMENT_MODE } from "@/core/lib";
 import { deckModule } from "./modules/deck";
 import { formatModule } from "./modules/format";
@@ -7,9 +6,7 @@ import { dataModule } from "./modules/data";
 import type { AppState } from "./AppState";
 import { collectionModule } from "./modules/collection";
 
-Vue.use(Vuex);
-
-export const store = new Store<AppState>({
+const store = createStore<AppState>({
 	modules: {
 		data: dataModule,
 
@@ -19,6 +16,7 @@ export const store = new Store<AppState>({
 	},
 	strict: DEVELOPMENT_MODE,
 });
+export default store;
 
 // Prepare for Vuex4 (https://next.vuex.vuejs.org/guide/typescript-support.html#simplifying-usestore-usage)
 export const useStore = (): typeof store => store;
